+++ web/src/components/dashboard/shared/index.ts
@@ -5,5 +5,6 @@
  */
 
 export * from './ChartContainer'
+export * from './ChartTooltip'
 export * from './DefaultStates'
 export * from './useDashboardState'